Sentiment/requirements_2.txt
Flask
Flask-Cors
python-dotenv
openai
huggingface-hub
scikit-learn
numpy
requests